Ass2-Classification/requirements.txt
torch>=1.13.0
torchvision>=0.14.0
timm>=0.9.0
pandas>=1.5.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.64.0
matplotlib>=3.5.0
